limit nerd fonts to symbols only; put font package reference in the correct place
The entire set of nerd fonts is flippin huge! Previously I tried to
install the symbols-only font, but it didn't work. That's because I put
the font package in `home.packages`, but I didn't hook it into the NixOS
font configuration.

It turns out the correct thing to do (in NixOS configuration at least)
is to include the font package in the `fonts.fonts` list (in 23.05 and
earlier - in later releases that will change to `fonts.packages`).

It looks like to install a font through Home Manager adding the font to
`home.packages` does work if you also set `fonts.fontconfig = true`