avoid segmentation fault from 1password in lock-session script
I was getting segmentation faults running `1password --lock`, but only
in the `lock-session` script when run by swayidle during the
`before-sleep` event (not when run directly). Based on the logs it looks
like the fault had to do with 1password failing to find an X display. So
added flags in the `lock-session` script to force 1password to run in
Wayland mode which seems to fix the problem.

I think this problem may have been preventing my computer from
suspending automatically.